+++ packages/opentelemetry-sdk-trace-base/test/common/BasicTracerProvider.test.ts
@@ -61,7 +61,7 @@ describe('BasicTracerProvider', () => {
   let envSource: Record<string, any>;
   let setGlobalPropagatorStub: sinon.SinonSpy<[TextMapPropagator], boolean>;
 
-  if (typeof process === 'undefined') {
+  if (global.process?.versions?.node === undefined) {
     envSource = globalThis as unknown as Record<string, any>;
   } else {
     envSource = process.env as Record<string, any>;